Resolve a dependency cycle by moving things from Jkind to Types
Once abstract kinds are added, Jkind must depend on Env.  But currently Env
depends on Subst, which depends on Jkind.  We break this cycle by making Subst
not depend on Jkind, moving the relevant functionality to Types (plus a little
to Jkind_types, where it belongs in modules that are already there).

The only real complication here is no_with_kinds_and_equal, which is needed by
the subst memo table but would need access to Env to implement real equality.
But its users (the memo table and printing) don't actually need equality, so
we move it to types and just document that it's shallow.