Eliminate dependency of `Subst` on `Jkind`
This moves enough stuff from `Jkind` to `Types` and `Jkind_types` so that
`Subst` no longer needs to depend on `Jkind`.

The only real complication here is `no_with_kinds_and_equal`, which is needed by
the subst memo table but would need access to `Env` to implement real equality
once we add abstract kinds.  But its users (the memo table and printing) don't
actually need equality, so we move it to types and just document that it's
shallow.